moved searchbar component to the components folder
