+++ app/components/searchbar.tsx
@@ -40,7 +40,7 @@ export default function Searchbar() {
   position: 'absolute', 
   top:'0', 
   zIndex: 100,
-  paddingBottom: 
+  paddingBottom: 50
   }}
   >
     <div
